third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/access-longest-css-property-expected.txt

The test parsing of the boundary value by checking the longest CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.body).webkitBorderBottomRightRadius is "5px"
PASS document.getElementById("target").style.webkitBorderBottomRightRadius is "5px"
PASS document.body.style.webkitBorderBottomRightRadius is ""
PASS successfullyParsed is true

TEST COMPLETE
